requirements.txt
python-telegram-bot>=13.0
python-mpd2>=3.0.0
tinytag>=1.5.0